refactor: change components folder name to lower case
