+++ src/features/image/index.ts
